added max python version, moved get_diams to utils, added option to download model from zenodo
