+++ setup.cfg
@@ -29,7 +29,7 @@ project_urls =
 packages = 
     napari_organoid_counter
 include_package_data = True
-python_requires = >=3.8
+python_requires = >=3.8 < 3.10
 setup_requires = setuptools_scm
 # add your package requirements here
 install_requires =
@@ -37,6 +37,7 @@ install_requires =
     napari-aicsimageio>=0.7.2
     torch>=1.13.1 
     torchvision>=0.14.1
+    zenodo_get==1.3.4
 
 #pytorch-lightning>=1.9.4
